MAINT: move `get_ordered_records` to `data_store` module
